bunch of begginings
added simple template, basic github tag lookup, and deployed environment
lookup functionality. 'environments' is kind of ambigous since we rely
on the env in the UNIX sense for some settings and overrides.

Also gave up on making this generic. Just going to rifle it at Socorro's
needs for a bit and we can refactor it later as necessary.