implement conditional request cache for github api
This patch removes user request-level cache for communicating with github and
introduces an indefinite cache on calls to the github api that invalidates keys
by making a conditional-request every time a datum is accessed.

the request-level cache may be restored in the future to improve page
speed, but this implementation should go a long way towards eliminating
problems related to github rate limiting, since our release tags never
change once they are made.